pyproject.toml
[tool.poetry]
name = "product-search-api"
version = "0.1.0"
description = ""
authors = ["OmarAlaraby <omar.alaraby23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
djangorestframework = "^3.16.0"
python-decouple = "^3.8"
psycopg2 = "^2.9.10"
drf-yasg = "^1.21.10"
langdetect = "^1.0.9"
deepl = "^1.22.0"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
